--- src/views/lComponent/wang-editor.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  <div>富文本</div>
-</template>
-
-<style scoped lang="less">
-
-</style>
